rename components; unite skills component in one component
